refactor(export): extract backend module
- Move _resolve_export_backend, _export_executorch_format,
  _EXPORT_FORMATS, _BACKEND_FORMATS from export/main.py to new
  export/_backend.py (helpers, not CLI orchestration)
- Update detr.py's two lazy imports to rfdetr.export._backend
- Update converter.py docstring cross-ref
- Update test imports + mock.patch target

---
